:hammer: REFACTOR (project): move project api to page
